moving getStartHour and getEndHour into Date extension
